added system tray icon, added blinking text
* Added system tray icon with a matching context menu from the software
* Added timer blinking text based on timer blink colours
* Changed timers so text at startup is timers startColour

+++ src/com/DBDKillerTimer/main/TimerBlink.java
@@ -8,5 +8,5 @@ public class TimerBlink {
     /** Time to set colour at */
     public int time;
     /** Frequency of text flash (ms), can be 0 to disable flashing text */
-    public long blinkFrequency;
+    public int blinkFrequency;
 }
